Extract the containers to common components
