+++ app.json
@@ -12,7 +12,7 @@
     "splash": {
       "image": "./assets/splash.png",
       "resizeMode": "contain",
-      "backgroundColor": "#ffffff"
+      "backgroundColor": "#F6F9FC"
     },
     "updates": {
       "fallbackToCacheTimeout": 0
