+++ src/screens/LoginScreen/index.js
@@ -1,20 +1,14 @@
 import React, { useContext } from "react";
-import { KeyboardAvoidingView, StyleSheet, ScrollView, Text, TouchableOpacity, View } from "react-native";
+import { KeyboardAvoidingView, StyleSheet, Text, TouchableOpacity  } from "react-native";
 import { useMutation } from "@apollo/react-hooks";
 
 import LoginMutation from "../../mutations/Login";
 import PlayerContext from "../../context/PlayerContext";
 
 import LoginForm from "../../components/LoginForm";
+import ScrollContainer from "../../components/ScrollContainer";
 
 const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "stretch",
-    justifyContent: "center",
-    padding: 5,
-  },
   header: {
     alignSelf: "center",
     fontSize: 40,
@@ -37,7 +31,7 @@ const LoginScreen = ({ navigation }) => {
   });
 
   return (
-    <ScrollView contentContainerStyle={styles.container} keyboardShouldPersistTaps="handled">
+    <ScrollContainer keyboardShouldPersistTaps="handled">
       <KeyboardAvoidingView behavior="position" keyboardVerticalOffset={20}>
         <Text style={styles.header}>Login</Text>
         <LoginForm onLogin={(auth) => login({ variables: auth})} />
@@ -45,7 +39,7 @@ const LoginScreen = ({ navigation }) => {
           <Text style={styles.linkButton}>Don&apos;t have an account?</Text>
         </TouchableOpacity>
       </KeyboardAvoidingView>
-    </ScrollView>
+    </ScrollContainer>
   );
 };
