+++ src/screens/RegistrationScreen/index.js
@@ -1,20 +1,14 @@
 import React, { useContext } from "react";
-import { KeyboardAvoidingView, StyleSheet, ScrollView, Text, TouchableOpacity, View } from "react-native";
+import { KeyboardAvoidingView, StyleSheet, Text, TouchableOpacity } from "react-native";
 import { useMutation } from "@apollo/react-hooks";
 
 import PlayerContext from "../../context/PlayerContext";
 import RegisterPlayerMutation from "../../mutations/RegisterPlayer";
 
 import RegistrationForm from "../../components/RegistrationForm";
+import ScrollContainer from "../../components/ScrollContainer";
 
 const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "stretch",
-    justifyContent: "center",
-    padding: 5,
-  },
   header: {
     alignSelf: "center",
     fontSize: 40,
@@ -37,7 +31,7 @@ const RegistrationScreen = ({ navigation }) => {
   });
 
   return (
-    <ScrollView contentContainerStyle={styles.container} keyboardShouldPersistTaps="handled">
+    <ScrollContainer keyboardShouldPersistTaps="handled">
       <KeyboardAvoidingView behavior="position" keyboardVerticalOffset={20}>
         <Text style={styles.header}>Player Registration</Text>
         <RegistrationForm
@@ -47,7 +41,7 @@ const RegistrationScreen = ({ navigation }) => {
           <Text style={styles.linkButton}>Already have an account?</Text>
         </TouchableOpacity>
       </KeyboardAvoidingView>
-    </ScrollView>
+    </ScrollContainer>
   );
 };
